complete refactor
