+++ index.html
@@ -10,6 +10,7 @@
 			href="https://fonts.googleapis.com/css2?family=Rubik:ital,wght@0,300..900;1,300..900&display=swap"
 			rel="stylesheet"
 		/>
+		<link rel="stylesheet" href="/styles.css" />
 		<script
 			src="https://kit.fontawesome.com/bd166a55fc.js"
 			crossorigin="anonymous"
@@ -17,22 +18,18 @@
 	</head>
 	<body>
 		<header>
-			<div id="header"></div>
+			<div id="header" class="bg-white border-b border-gray-200"></div>
 		</header>
-		<main class="grid h-screen place-items-center container">
-			<div id="tutorial-wrapper">
-				<div id="tutorial-header"></div>
-				<div
-					id="tutorial-content"
-					class="mx-auto flex max-w-7xl items-center justify-between"
-				></div>
-				<div id="tutorial-footer"></div>
-			</div>
+		<main class="grid place-items-center">
+			<div id="sub-header" class="py-2"></div>
+			<div id="app"></div>
 		</main>
 		<footer>
-			<div id="footer"></div>
+			<div id="footer" class="bg-white border-t border-gray-200"></div>
 		</footer>
-		<script type="module" src="/src/components/ui/header.ts"></script>
-		<script type="module" src="/src/app.ts"></script>
+		<script type="module" src="/src/components/header.ts"></script>
+		<script type="module" src="/src/components/subheader.ts"></script>
+		<script type="module" src="/src/components/main.ts"></script>
+		<script type="module" src="/src/components/footer.ts"></script>
 	</body>
 </html>
