+++ package.json
@@ -11,7 +11,7 @@
   "devDependencies": {
     "@types/node": "^20.11.24",
     "autoprefixer": "^10.4.18",
-    "postcss": "^8.4.35",
+    "postcss": "^8.4.38",
     "tailwindcss": "^3.4.1",
     "ts-node": "^10.9.2",
     "typescript": "^5.2.2",
