--- src/app.ts
@@ -1,4 +1,0 @@
-import { initializeTutorial } from "./components/ui/main";
-import "./style.css";
-
-document.addEventListener("DOMContentLoaded", initializeTutorial);
