--- src/components/tutorial/content/conclusionContent.ts
@@ -1,9 +1,0 @@
-interface conclusionContent {
-	content: string;
-}
-
-export const conclusionContent: conclusionContent[] = [
-	{ content: `<p>Content part 1</p>` },
-	{ content: `<p>Content part 2</p>` },
-	{ content: `<p>Content part 3</p>` },
-];
