--- src/components/tutorial/content/introContent.ts
@@ -1,223 +1,0 @@
-import "../../../style.css";
-
-interface IntroContent {
-	content: string;
-}
-
-export const introContent: IntroContent[] = [
-	{
-		content: `                   
-        <h2>
-            Common Approaches to Software Development
-        </h2>
-        <p>
-            Developing software is an art and science that involves various approaches, each with its strengths and optimal use cases. Let's take a look at three common approaches used solo developers and both small and large teams:
-        </p>
-        <ul>
-            <li>
-                <strong>Waterfall 🌊:</strong> Waterfall is the traditional phase-oriented methodology. It's like a step-by-step recipe where you complete one step before moving on to the next. Think of it as a waterfall flowing downstream, where each phase cascades into the next, from requirements to maintenance.
-            </li>
-            <li>
-                <strong>Agile 🌀:</strong> Agile is a flexible, practice-oriented approach. Picture a series of loops or cycles, where planning, development, and evaluation happen in quick, repetitive rounds. It's like preparing dishes in a professional kitchen—fast, adaptable, and customer-focused, with continuous tasting and improvement.
-            </li>
-            <li>
-                <strong>Extreme Programming (XP) ⚙️:</strong> XP is a subset of Agile that emphasizes technical best practices and high-quality results through a set of defined practices. Imagine a set of gears working together seamlessly or a team sport where close collaboration and quick passes lead to a score. It's all about constant communication, feedback, and refinement.
-            </li>
-        </ul>`,
-	},
-	{
-		content: `
-        <h2>
-            Waterfall Principles & Characteristics
-        </h2>
-        <p>
-            The Waterfall methodology, one of the earliest software development methods, is a linear and sequential approach where each phase of the development process flows steadily downwards, like a waterfall, through the phases of conception, initiation, analysis, design, construction, testing, deployment, and maintenance.
-        </p>
-        <h3>
-            Principles
-        </h3>
-        <ul>
-            <li>
-            <strong>➡️Sequential Process Flow:</strong> The Waterfall model is characterized by a series of steps that flow in a linear sequence. Each phase must be completed before moving on to the next, with little to no overlap between phases.
-            </li>
-            <li>
-            <strong>🛑Completion Before Progression:</strong> No phase begins until the prior phase is complete, ensuring thoroughness and accuracy at each step. This principle reinforces discipline in project management and execution.
-            </li>
-            <li>
-            <strong>📝Documentation and Planning Emphasis:</strong> Detailed documentation is a cornerstone of the Waterfall methodology. Each phase is well-documented, providing a clear set of deliverables that guide the next phase. This rigorous documentation facilitates clear understanding and reference throughout the project lifecycle.
-            </li>
-        </ul>
-        <h3>
-            Characteristics
-        </h3>
-        <ul>
-            <li><strong>1️⃣Highly Structured:</strong> The Waterfall model's strength lies in its structured design, making it predictable and easy to understand. Each phase has specific deliverables and a review process, contributing to a straightforward management and execution process.
-            </li>
-            <li>
-            <strong>2️⃣Predictability and Control:</strong> Due to its linear progression, the Waterfall model allows for enhanced control over the project timeline, budget, and resources. This predictability makes it easier for project managers to allocate resources effectively.
-            </li>
-            <li>
-            <strong>3️⃣Best Suited for Well-Defined Projects:</strong> Projects with clear objectives, well-defined requirements, and a low risk of change are ideal candidates for the Waterfall methodology. Its structured nature makes it less flexible, thus less suited for projects where requirements are expected to evolve.
-            </li>
-        </ul>`,
-	},
-	{
-		content: `
-        <h2>
-            Agile Principles & Characteristics
-        </h2>
-        <p>
-            Agile methodology is a set of principles for software development under which requirements and solutions evolve through the collaborative effort of self-organizing and cross-functional teams. It advocates adaptive planning, evolutionary development, early delivery, and continual improvement, and it encourages flexible responses to change.
-        </p>
-        <h3>
-            Principles
-        </h3>
-        <ul>
-            <li>
-            <strong>🧑‍🤝‍🧑Individuals and Interactions Over Processes and Tools:</strong> Agile prioritizes human communication and feedback, valuing the dynamic interaction of team members over rigid adherence to tools and processes.
-            </li>
-            <li>
-            <strong>👨‍💻Working Software Over Comprehensive Documentation:</strong> Agile values the functionality of the final product more than detailed documentation. While documentation is not disregarded, it is streamlined to what is necessary to deliver the next piece of functional software.
-            </li>
-            <li>
-            <strong>🤝Customer Collaboration Over Contract Negotiation:</strong> Agile involves customers in the development process, preferring collaborative decision-making over sticking strictly to the terms of a contract. This principle ensures that the product evolves according to actual user needs.
-            </li>
-            <li>
-            <strong>🛞Responding to Change Over Following a Plan:</strong> Agile teams are expected to be responsive to change, even late in the development process. This flexibility ensures the product remains relevant and competitive in a changing market environment.
-            </li>
-        </ul>
-        <h3>
-            Characteristics
-        </h3>
-        <ul>
-            <li><strong>1️⃣Flexibility and Adaptiveness: :</strong> Agile teams can quickly pivot in response to feedback or changing requirements. This adaptiveness reduces the risk of developing an outdated or irrelevant product.
-            </li>
-            <li>
-            <strong>2️⃣Emphasis on Direct Communication:</strong> Regular, face-to-face communication among team members and between stakeholders and the development team is a cornerstone of Agile, fostering clarity and reducing the likelihood of misunderstandings.
-            </li>
-            <li>
-            <strong>3️⃣Frequent Delivery of Functional Software:</strong> Agile focuses on the continuous delivery of small increments of working software, allowing teams to demonstrate progress and adjust as needed.
-            </li>
-            <li>
-            <strong>4️⃣Customer Involvement:</strong> Agile methodology encourages ongoing customer involvement to ensure the product meets their needs and expectations, leading to higher customer satisfaction.
-            </li>
-        </ul>`,
-	},
-	{
-		content: `
-        <h2>
-            Extreme Programming (XP) Principles & Characteristics
-        </h2>
-        <p>
-            Extreme Programming (XP) is an agile software development framework that aims to produce higher quality software, and higher quality of life for the development team. XP is characterized by its short development cycles, incremental planning approach, continuous feedback, and—most notably—its emphasis on technical excellence and good design.
-        </p>
-        <h3>
-            Principles
-        </h3>
-        <ul>
-            <li>
-            <strong>🧑‍🤝‍🧑Close Collaboration Between Developers and Customers:</strong> Agile prioritizes human communication and feedback, valuing the dynamic interaction of team members over rigid adherence to tools and processes.
-            </li>
-            <li>
-            <strong>👩‍🎨Simplicity in Design:</strong> XP promotes the idea that systems should be designed as simply as possible. This simplicity allows for easier changes in the future and avoids unnecessary complexity that can lead to errors or increased maintenance costs.
-            </li>
-            <li>
-            <strong>🔁Frequent Feedback:</strong> Regular feedback from the customer, and within the development team, is crucial in XP. This feedback loop helps the team stay aligned with customer needs and adjust practices as the project evolves.
-            </li>
-            <li>
-            <strong>🤝Team Empowerment:</strong> XP believes that every team member should have a voice in the development process. This empowerment encourages innovation and shared ownership of the project, leading to more creative solutions and a stronger, more cohesive team.
-            </li>
-        </ul>
-        <h3>
-            Characteristics
-        </h3>
-        <ul>
-            <li><strong>1️⃣High Collaboration and Iteration:</strong> XP involves customers in the development process to a high degree, with frequent releases and iterations that ensure the product is developing in alignment with customer needs.
-            </li>
-            <li>
-            <strong>2️⃣Emphasis on Technical Excellence:</strong> XP teams strive for excellence in coding and testing practices. Continuous integration, test-driven development, and pair programming are key practices that enhance product quality and team productivity.
-            </li>
-            <li>
-            <strong>3️⃣Adaptability to Change:</strong> XP is designed to accommodate changes in customer requirements, even late in the development cycle. This flexibility ensures the final product remains relevant and valuable to the customer.
-            </li>
-        </ul>`,
-	},
-	{
-		content: `
-        <h2>
-            So, what method is best for your project?
-        </h2>
-        <p>
-            The short answer is: <strong>it depends</strong>.
-        </p>
-        <div class="overflow-x-auto shadow-md sm:rounded-lg mb-10">
-            <table class="w-full text-sm text-left">
-                <thead class="text-xs text-gray-700 uppercase bg-gray-50">
-                    <tr>
-                        <th scope="col" class="px-6 py-3">
-                            <div class="flex items-center">
-                            🌊 Waterfall
-                            </div>
-                        </th>
-                        <th scope="col" class="px-6 py-3">
-                            <div class="flex items-center">
-                                🌀 Agile
-                            </div>
-                        </th>
-                        <th scope="col" class="px-6 py-3">
-                            <div class="flex items-center">
-                                ⚙️ XP
-                            </div>
-                        </th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <tr class="bg-white dark:bg-gray-800 dark:border-gray-700">
-                        <td class="px-6 py-4">
-                            Requirements are well-understood and unlikely to change.
-                        </td>
-                        <td class="px-6 py-4">
-                            Requirements are expected to change or are not fully understood at the outset.
-                        </td>
-                        <td class="px-6 py-4">
-                            The project scope is expected to change or is not fully defined at the outset.
-                        </td>
-                    </tr>
-                    <tr class="bg-white dark:bg-gray-800 dark:border-gray-700">
-                        <td class="px-6 py-4">
-                            The project scope is clear, and the product definition is stable.
-                        </td>
-                        <td class="px-6 py-4">
-                            The project benefits from frequent reviews and adaptations, allowing for rapid pivots based on user feedback or market trends.
-                        </td>
-                        <td class="px-6 py-4">
-                            The risk of technological change is high, and the project needs to adapt rapidly.
-                        </td>
-                    </tr>
-                    <tr class="bg-white dark:bg-gray-800">
-                        <td class="px-6 py-4">
-                            There is a need for strict regulatory documentation and compliance.
-                        </td>
-                        <td class="px-6 py-4">
-                            Innovation, speed, and flexibility are more critical than predictability and control.
-                        </td>
-                        <td class="px-6 py-4">
-                            There is a need for frequent delivery of functional software to accommodate fast-paced environments.
-                        </td>
-                    </tr>
-                    <tr class="bg-white dark:bg-gray-800">
-                        <td class="px-6 py-4">
-                            The technology being used is well-understood, and there are no exploratory elements involved.
-                        </td>
-                        <td class="px-6 py-4">
-                            The project environment supports collaboration and communication among cross-functional teams.
-                        </td>
-                        <td class="px-6 py-4">
-                            A high level of customer involvement is possible and desired to ensure the product meets real user needs.
-                        </td>
-                    </tr>
-                </tbody>
-            </table>
-    </div>`,
-	},
-];
