--- src/components/tutorial/content/partFiveContent.ts
@@ -1,9 +1,0 @@
-interface PartFiveContent {
-	content: string;
-}
-
-export const partFiveContent: PartFiveContent[] = [
-	{ content: `<p>Content part 1</p>` },
-	{ content: `<p>Content part 2</p>` },
-	{ content: `<p>Content part 3</p>` },
-];
