--- src/components/tutorial/content/partFourContent.ts
@@ -1,9 +1,0 @@
-interface PartFourContent {
-	content: string;
-}
-
-export const partFourContent: PartFourContent[] = [
-	{ content: `<p>Content part 1</p>` },
-	{ content: `<p>Content part 2</p>` },
-	{ content: `<p>Content part 3</p>` },
-];
