--- src/components/tutorial/content/partOneContent.ts
@@ -1,61 +1,0 @@
-interface PartOneContent {
-	content: string;
-}
-
-export const partOneContent: PartOneContent[] = [
-	{
-		content: `
-			<h2>Where to begin?</h2>
-			<p>Every great software project begins with a solid foundation. This initial phase, although called by different names in various methodologies, shares a common goal: <strong>to lay down the groundwork for successful development.</strong></p> 
-			<p>In this part of the tutorial, we'll explore how the Waterfall, Agile, and Extreme Programming (XP) methodologies approach this crucial stage.</p>`,
-	},
-	{
-		content: `
-			<h2>Getting Started with Waterfall</h2>
-			<p>In Waterfall, the journey begins with clarifying your project's goals during the <strong>Requirements</strong> phase. It's about taking your idea and outlining exactly what you want your project to achieve, considering both the features it will have and how it should perform.</p>
-			<h3>Action Plan:</h3>
-			<ol>
-				<li>
-					<strong>Define Your Project Goals:</strong> Start by writing down what problem your project solves or what it aims to achieve. Think about who will use it and in what context.</li>
-				<li>
-					<strong>Specify Functional and Non-Functional Requirements:</strong> Use a simple document to list all features (functional requirements) and performance expectations like speed, security, and usability (non-functional requirements). Tools like Notion or Google Docs can keep things organized.</li>
-				<li>
-					<strong>Sketch Use Cases:</strong> Draft scenarios showing how your project will be used. This helps in visualizing the user's journey and ensures the project is user-centric. Draw.io or simple sketches on paper can be effective.</li>
-			</ol>`,
-	},
-	{
-		content: `
-			<h2>Getting Started with Agile</h2>
-			<p>Agile methodologies introduce the project through the <strong>Concept</strong> or <strong>Inception</strong> phase. Agile methodologies like Scrum or Kanban emphasize flexibility and iterative development. The Concept or Inception phase helps you set a broad vision and outline the scope without getting bogged down in details.</p>
-			<h3>Action Plan:</h3>
-			<ol>
-				<li>
-					<strong>Craft Your Vision and Goals:</strong> Use a tool like Trello to brainstorm and categorize ideas about what your project will do. Think big picture and be ready to adapt.</li>
-				<li>
-					<strong>Outline High-Level Requirements with User Stories:</strong> Write down user stories in a notebook or a digital tool like Jira. Each story should focus on a specific user need or goal your project addresses.</li>
-				<li>
-					<strong>Value Proposition Canvas:</strong> Even if you're working solo, understanding the value your project delivers is crucial. Sketch out a Value Proposition Canvas to clarify the benefits and features.</li>
-			</ol>`,
-	},
-	{
-		content: `
-			<h2>Getting Started with XP</h2>
-			<p>Extreme Programming (XP) starts with the <strong>Planning</strong> phase. XP focuses on continuous feedback and simplicity. This phase is about laying out the groundwork with user stories and a basic plan for your project's development cycle.</p>
-			<h3>Action Plan:</h3>
-			<ol>
-				<li>
-					<strong>Create User Stories:</strong> Capture the functionality of your project from a user's perspective using index cards or a digital board like GitHub Projects. This makes it easier to prioritize tasks.</li>
-				<li>
-					<strong>Plan Releases:</strong> Decide what features make up your first version. Tools like Trello can help you organize stories into sprints or phases of development.</li>
-				<li>
-					<strong>System Metaphors:</strong> Develop a simple metaphor or analogy that describes your project's overall system. This helps in keeping the design and development focused and coherent.</li>
-			</ol>`,
-	},
-	{
-		content: `
-			<h2>Understanding the Foundation</h2>
-			<p>In all three methodologies, the initial phase is crucial for setting the direction and ensuring that the project's objectives are clear. Whether it's called <strong>Requirements</strong>, <strong>Concept/Inception</strong>, or <strong>Planning</strong>,  the goal is to start with a clear understanding of what you want to build and for whom, setting the stage for a successful project.For new or solo developers, these initial stages are about turning your vision into a concrete plan that guides your development process.</p>
-
-			<p>As we proceed through this tutorial, we'll delve deeper into how these methodologies build upon their initial phases to manage software development processes. Each approach has its strengths and challenges, but they all start by laying a solid foundation for success. Remember, these methodologies are not strict rules but frameworks to help organize your thoughts and work. Feel free to adapt or combine them as needed to suit your project and working style best.</p>`,
-	},
-];
