--- src/components/tutorial/content/partThreeContent.ts
@@ -1,9 +1,0 @@
-interface PartThreeContent {
-	content: string;
-}
-
-export const partThreeContent: PartThreeContent[] = [
-	{ content: `<p>Content part 1</p>` },
-	{ content: `<p>Content part 2</p>` },
-	{ content: `<p>Content part 3</p>` },
-];
