--- src/components/tutorial/content/partTwoContent.ts
@@ -1,71 +1,0 @@
-interface PartTwoContent {
-	content: string;
-}
-
-export const partTwoContent: PartTwoContent[] = [
-	{
-		content: `
-		<h2>From Vision to Blueprint</h2>
-		<p>After establishing a solid foundation for our software project through careful planning and requirements gathering, we move into one of the most critical stages: <strong>Design</strong>. In this phase, we transition from what we want our software to achieve to how we will build it. This part of our journey is where creativity meets practicality, and visions start taking a tangible form.</p>
-		<p>Design isn't just about aesthetics; it's about creating a blueprint that guides the development process. It involves making key decisions on the architecture, technologies, and frameworks that will be used to bring the project to life. Each methodology approaches design with its unique perspective and tools, but the goal remains the same: <strong>to devise a robust, scalable, and efficient system that meets the project's requirements and is adaptable to future needs.</strong></p>
-		<p>As we delve into each methodology, remember that the design phase is your opportunity to solve problems creatively and set your project on the path to success. Let's explore how you can make the most of this phase, regardless of the methodology you choose.</p>`,
-	},
-	{
-		content: `<h2>Waterfall's Design Phase</h2>
-			<p>The <strong>Design</strong> phase in Waterfall is about taking your documented requirements and turning them into a detailed software design. This phase lays out the architecture, components, interfaces, and data for your project in a way that addresses all identified requirements.</p>
-			<h3>Mastering Waterfall's Design Phase:</h3>
-			<ol>
-				<li>
-					<strong>Create Architectural Diagrams:</strong> Use tools like Lucidchart or Draw.io to map out the high-level architecture of your project, showing how different parts of your system will interact.</li>
-				<li>
-					<strong>Detail Component Design:</strong> For each component or module, draft detailed designs that describe the logic, data structures, and algorithms needed. Simplify complex problems by breaking them down into smaller, manageable parts.</li>
-				<li>
-					<strong>Interface Definitions:</strong> Clearly define how different parts of your application will communicate with each other and with any external systems. Document APIs, data formats, and protocols.
-					</li>
-				<li>
-					<strong>Review and Refine:</strong> Developers can benefit from self-review or peer review through developer communities like Stack Overflow or GitHub to ensure the design meets the project goals and is feasible with available resources.
-				</li>
-			</ol>`,
-	},
-	{
-		content: `<h2>Sprint Planning with Agile</h2>
-			<p>Agile methodologies, particularly Scrum, use <strong>Sprint Planning</strong> to prepare for a development cycle (sprint). This is where you decide what work will be done during the sprint based on your project's backlog (prioritized list of user stories).</p>
-			<h3>Mastering Agile's Sprint Planning:</h3>
-			<ol>
-				<li>
-					<strong>Prioritize User Stories:</strong> Review your backlog and select the most important or urgent user stories for the upcoming sprint. Tools like Jira or Asana can help manage your backlog.</li>
-				<li>
-					<strong>Task Breakdown:</strong> Break each selected user story into tasks and estimate their complexity or time required. This can be done using Trello boards or GitHub Projects.</li>
-				<li>
-					<strong>Set Sprint Goals:</strong> Define a clear goal for the sprint based on the selected user stories. This goal should reflect a tangible outcome that brings value to users.
-					</li>
-				<li>
-					<strong>Prepare the Sprint Board:</strong> Organize tasks on a sprint board, digital or physical, to track progress during the sprint. Ensure all tasks are actionable and clearly defined.
-				</li>
-			</ol>`,
-	},
-	{
-		content: `<h2>Designing in XP</h2>
-			<p><strong>Designing</strong> in XP emphasizes simplicity and the principle of doing only what's necessary at the current stage. The design is continuously refined based on feedback and new insights as development progresses.</p>
-			<h3>Mastering XP's Designing Phase:</h3>
-			<ol>
-				<li>
-					<strong>Prioritize User Stories:</strong> Review your backlog and select the most important or urgent user stories for the upcoming sprint. Tools like Jira or Asana can help manage your backlog.</li>
-				<li>
-					<strong>Use System Metaphors:</strong> Develop and refine system metaphors that guide the design and communication about the system. This helps maintain coherence in your project.</li>
-				<li>
-					<strong>Continuous Refactoring:</strong> Be prepared to refine and optimize the design as you develop and learn more about the problem space. Tools like Refactoring Guru provide guidance on refactoring patterns.
-					</li>
-				<li>
-					<strong>Test-Driven Development (TDD):</strong> Use TDD as a design tool. Writing tests before code not only ensures your code works as intended but also helps clarify the design and interfaces of your components.
-				</li>
-			</ol>`,
-	},
-	{
-		content: `
-		<h2>User-Focused Design is <i>KEY</i>!</h2>
-		<p>The design phase stands at the intersection of planning and innovation. It's where the project's vision transforms into a technical blueprint. This blueprint might take the form of comprehensive diagrams in the Waterfall method, flexible sprint plans in Agile, or the pursuit of simplicity in XP. The essence of this phase is not to set constraints through rigid planning but to chart a course for crafting solutions that resonate with both your objectives and your users' needs.</p>
-
-		<p>A pivotal aspect of successful design is its user-centric focus. Designing with the user in mind ensures that your project is not just technically sound but also valuable and intuitive to those who matter most—the end users. Keep your designs adaptable, allowing them to evolve as you gain deeper insights into user requirements and behaviors. Remember, the true measure of your project's success is in its ability to meet user expectations and solve real-world problems effectively.</p>`,
-	},
-];
