--- src/components/tutorial/content/tutorialContent.ts
@@ -1,7 +1,0 @@
-export { loadIntro } from "../../../tutorial/intro";
-export { loadPartOne } from "../tutorial/part1";
-export { loadPartTwo } from "../tutorial/part2";
-export { loadPartThree } from "../tutorial/part3";
-export { loadPartFour } from "../tutorial/part4";
-export { loadPartFive } from "../tutorial/part5";
-export { loadConclusion } from "../conclusion";
