--- src/components/tutorial/content/welcomeContent.ts
@@ -1,72 +1,0 @@
-interface WelcomeContent {
-	content: string;
-}
-
-export const welcomeContent: WelcomeContent[] = [
-	{
-		content: `                   
-        <h2>
-            Welcome, software developers!
-        </h2>
-        <p>
-            Whether you're starting your journey in the tech world or looking to solidify your foundational knowledge, you've come to the right place. This free tutorial has been crafted to guide you through the essentials of planning and executing software projects, tailored specifically for solo developers and small teams.
-        </p>
-        <p>
-            Understanding the methodology behind project planning and execution is as crucial as coding itself. It's not just about writing code; it's about crafting solutions that are efficient, scalable, and, above all, impactful.
-        </p>`,
-	},
-	{
-		content: `
-        <h2>
-            What You'll Learn:
-        </h2>
-        <ul>
-            <li">
-                <strong>The Essentials of Software Project Planning:</strong> Dive into the art and science of turning ideas into actionable project plans. Learn to outline your project's scope, define its objectives, and set achievable milestones.
-            </li>
-            <li>
-                <strong>Execution Strategies for Solo Devs and Small Teams:</strong> Discover the tools, techniques, and practices that can make or break software projects in lean environments. Find out how to keep your project on track even when resources are limited.
-            </li>
-            <li>
-                <strong>Navigating Through Development Approaches:</strong> Unravel the mysteries behind the three most popular software development methodologies today:
-                <ul>
-                    <li>
-                        <strong>Waterfall:</strong> A sequential, phase-based approach that's straightforward and disciplined.
-                    </li>
-                    <li>
-                        <strong>Agile:</strong> A dynamic and iterative method that embraces change and encourages continuous feedback.
-                    </li>
-                    <li>
-                        <strong>Extreme Programming (XP):</strong> A rigorous yet flexible practice focusing on excellence in software engineering.
-                    </li>
-                </ul>
-            </li>
-        </ul>`,
-	},
-	{
-		content: `
-        <h2>
-            Who Is This For?
-        </h2>
-        <p>
-            This tutorial is tailor-made for:
-        </p>
-        <ul class="list-disc ml-10 mb-4">
-            <li>
-                Aspiring developers embarking on their software development journey.
-            </li>
-            <li>
-                Experienced coders seeking to understand project planning and execution better.
-            </li>
-            <li>
-                Anyone curious about the methodologies that have shaped the software industry.
-            </li>
-        </ul>
-        <h3>
-            Let's Get Started!
-        </h3>
-        <p>
-            Ready to dive in? Click "<strong>Start Learning</strong>" to begin your exploration into the methodologies that power software development across the globe. Your path to becoming a proficient developer in planning and executing software projects begins now!
-        </p>`,
-	},
-];
