+++ src/interfaces.ts
@@ -1,0 +1,13 @@
+export interface ContentItem {
+	content: string;
+}
+export interface TutorialPart {
+	id: string;
+	title: string;
+	contentText: ContentItem[];
+	contentClass: string;
+	buttonText?: string;
+	shouldTutNavShow: boolean;
+	headerText?: string;
+	headerImage?: string;
+}
