+++ src/styles.css
@@ -84,11 +84,36 @@
 	}
 }
 
-#app {
-	max-width: 1280px;
-	margin: 0 auto;
-	padding: 2rem;
-	text-align: center;
+html,
+body {
+	height: 100%;
+	margin: 0;
+	padding: 0;
+	overflow: hidden;
+}
+
+header {
+	height: 90px;
+	width: 100%;
+	position: fixed;
+	top: 0;
+	z-index: 1000;
+}
+
+main {
+	position: fixed;
+	top: 110px;
+	bottom: 40px;
+	width: 100%;
+	overflow-y: scroll;
+}
+
+footer {
+	position: fixed;
+	bottom: 0;
+	width: 100%;
+	height: 40px;
+	z-index: 1000;
 }
 
 p,
@@ -128,20 +153,20 @@ ul > li > ul > li:first-of-type {
 	margin-top: 10px;
 }
 
-@media (prefers-color-scheme: light) {
-	:root {
-		color: #213547;
-		background-color: #ffffff;
-	}
-	a:hover {
-		color: #747bff;
-	}
-	button {
-		background-color: #f9f9f9;
-	}
-}
-
-#welcome {
+#content-display {
 	transition: opacity 0.2s ease-in-out;
 	opacity: 1;
+}
+
+#app {
+	max-width: 1280px;
+	margin: 0 auto;
+	padding: 2rem;
+	text-align: center;
+}
+
+#sub-header {
+	position: fixed;
+	top: 90px;
+	width: 100%;
 }
